chore: organize default layout
