--- src/layouts/rooms.vue
@@ -1,15 +1,0 @@
-<template>
-  <v-app>
-    <the-appbar />
-    <v-content>
-      <nuxt />
-    </v-content>
-    <the-bottom-nav />
-  </v-app>
-</template>
-
-<script>
-export default {}
-</script>
-
-<style></style>
