+++ src/layouts/talk.vue
@@ -1,0 +1,22 @@
+<template>
+  <v-app>
+    <the-appbar />
+    <v-content>
+      <nuxt />
+    </v-content>
+    <the-messenger />
+  </v-app>
+</template>
+
+<script>
+import TheAppbar from '~/components/organisms/TheAppbar'
+import TheMessenger from '~/components/organisms/TheMessenger'
+export default {
+  components: {
+    TheAppbar,
+    TheMessenger
+  }
+}
+</script>
+
+<style></style>
