+++ src/pages/index.vue
@@ -1,21 +1,9 @@
 <template>
-  <div>
-    <v-message />
-  </div>
+  <div />
 </template>
 
 <script>
-import { mixin } from '~/mixins/navbar'
-import VMessage from '~/components/organisms/VMessage'
-export default {
-  components: {
-    VMessage
-  },
-  mixins: [mixin],
-  data: () => ({
-    title: 'Talk'
-  })
-}
+export default {}
 </script>
 
 <style></style>
