+++ src/store/message/state.js
@@ -1,3 +1,11 @@
 export default () => ({
   messages: []
 })
+// message:{
+//   uuid:String,
+//   userid:String,
+//   text:String,
+//   tag:Number,
+//   createdAt:Date,
+//   deletedAt:Date
+// }
