+++ src/store/user/state.js
@@ -1,3 +1,9 @@
 export default () => ({
   user: null
 })
+// user:{
+//   uuid:String,
+//   name:String,
+//   createdAt:Date,
+//   deletedAt:Date
+// }
